[WIP] Event Streaming
This is a WIP for streaming events using a GenServer connected to a long
lived HTTP connection.

This commit also includes a small code refactor, changing the
V1_41Adapter to the DefaultAdapter (as I only plan to support one
version of Docker at a time).

It also adds the `Docker.Container.prune` function that does the same as
`docker containers prune`.